refactor(migrations): squash all 28 migrations into single baseline
We have no external users and the hosted database already has all
tables and columns. The 28 incremental migrations added complexity
without value — any new deploy just needs to know "you're at head."

This replaces the approach in PR #751 (smart migration script) with
the simpler strategy of squashing to a no-op baseline revision.
Future migrations will build from 0001 as the starting point.
